getting rid of fragments and fixing bug of stop method
