+++ .idea/misc.xml
@@ -8,6 +8,7 @@
         <entry key="..\:/Users/jason/AndroidStudioProjects/MyFocusApp/app/src/main/res/drawable/ic_timer.xml" value="0.1505" />
         <entry key="..\:/Users/jason/AndroidStudioProjects/MyFocusApp/app/src/main/res/layout/activity_main.xml" value="0.22826086956521738" />
         <entry key="..\:/Users/jason/AndroidStudioProjects/MyFocusApp/app/src/main/res/layout/activity_settings.xml" value="0.192481884057971" />
+        <entry key="..\:/Users/jason/AndroidStudioProjects/MyFocusApp/app/src/main/res/layout/activity_timer.xml" value="0.192481884057971" />
         <entry key="..\:/Users/jason/AndroidStudioProjects/MyFocusApp/app/src/main/res/layout/content_main.xml" value="0.1" />
         <entry key="..\:/Users/jason/AndroidStudioProjects/MyFocusApp/app/src/main/res/layout/fragment_home.xml" value="0.264" />
         <entry key="..\:/Users/jason/AndroidStudioProjects/MyFocusApp/app/src/main/res/layout/fragment_setting.xml" value="0.10989583333333333" />
