+++ app/src/main/AndroidManifest.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     package="com.example.myfocusapp">
 
     <uses-permission android:name="android.permission.SCHEDULE_EXACT_ALARM" />
 
-    <application
+    <application android:debuggable="true"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -12,10 +13,8 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.MyFocusApp">
-        <activity
-            android:name=".SettingsActivity"
-            android:exported="false" />
+        android:theme="@style/Theme.MyFocusApp"
+        tools:ignore="HardcodedDebugMode">
 
         <receiver
             android:name=".TimerNotificationActionReceiver"
